+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - PR #2585 ORC & Parquet Readers: Remove millisecond timestamp restriction
 - PR #2559 Add Series.tolist()
 - PR #2653 Add Java bindings for rolling window operations
+- PR #2674 Add __contains__ for Index/Series/Column
 
 ## Improvements